requirements.txt
loguru
https://github.com/KurimuzonAkuma/pyrogram/archive/v2.1.28.zip
PyYAML
TgCrypto
python-dotenv
httpx
tenacity